move billboard_label actor to actor.py
